back/requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
pymongo==4.15.3
beanie==2.0.0
motor==3.7.1
python-dotenv==1.2.1
pydantic-settings==2.7.0
pandas
numpy
httpx
